common query paramters
