+++ api_swedeb/api/dummy_data/dummy_speech.py
@@ -1,3 +1,4 @@
+from api_swedeb.api.utils.common_params import CommonQueryParams
 from api_swedeb.schemas.speeches_schema import SpeechesResult, SpeechesResultItem
 from api_swedeb.schemas.speech_text_schema import SpeechesTextResultItem
 from fastapi import HTTPException
@@ -12,12 +13,12 @@ def get_speech_by_id(id: str):
     )
 
 
-def get_speeches():
+def get_speeches(commons: CommonQueryParams):
     return SpeechesResult(
         speech_list=[
             SpeechesResultItem(
                 speaker_column="En talare",
-                year_column="1960",
+                year_column= commons.from_year if commons.from_year else "1960",
                 gender_column="M",
                 source_column="www.riksdagen.se",
                 speech_id_column="1",
